Student's t Deviances

Student's t Deviances

dev_student(x, mean = 0, sd = 1, theta = 0, res = FALSE)

Arguments

x

A numeric vector of values.

mean

A numeric vector of the means.

sd

A non-negative numeric vector of the standard deviations.

theta

A non-negative numeric vector of the dispersion for the mixture models (student, gamma-Poisson and beta-binomial).

res

A flag specifying whether to return the deviance residual as opposed to the deviance.

Value

An numeric vector of the corresponding deviances or deviance residuals.

See also

Examples

dev_student(c(1, 3.5, 4), 3)
#> [1] 4.00 0.25 1.00