Parameter Descriptions

Parameter Descriptions

Arguments

x

An object.

...

Other arguments passed to methods.

nas

A flag specifying whether to also fill missing values.

n

A non-negative whole number of the number of random samples to generate.

conf_level

A numeric scalar between 0 and 1 specifying the confidence level.

value

A scalar of the value to replace values with.

directional

A flag specifying whether probabilities less than 0.5 should be returned as negative values.

res

A flag specifying whether to return the deviance residual as opposed to the deviance.

mean

A numeric vector of the means.

sd

A non-negative numeric vector of the standard deviations.

meanlog

A numeric vector of the means on the log scale.

sdlog

A non-negative numeric vector of the standard deviations on the log scale.

size

A non-negative whole numeric vector of the number of trials.

prob

A numeric vector of values between 0 and 1 of the probability of success.

lambda

A non-negative numeric vector of means.

theta

A non-negative numeric vector of the dispersion for the mixture models (student, gamma-Poisson and beta-binomial).

shape

A non-negative numeric vector of shape.

rate

A non-negative numeric vector of rate.

type

A string of the residual type. 'raw' for raw residuals 'dev' for deviance residuals and 'data' for the data.

na_rm

A flag specifying whether to remove missing values.

threshold

A number of the threshold value.

simulate

A flag specifying whether to simulate residuals.

log

A flag specifying whether to return the log-transformed value.

lower.tail

A flag specifying whether to return the lower or upper tail of the distribution.

q

A vector of quantiles.

p

A vector of probabilities.

sd_mult

A non-negative multiplier on the standard deviation of the distribution.

alpha

The first shape parameter of the Beta distribution.

beta

The second shape parameter of the Beta distribution.