Student's t Random Samples

Student's t Random Samples

ran_student(n = 1, mean = 0, sd = 1, theta = 0)

Arguments

n

A non-negative whole number of the number of random samples to generate.

mean

A numeric vector of the means.

sd

A non-negative numeric vector of the standard deviations.

theta

A non-negative numeric vector of the dispersion for the mixture models (student, gamma-Poisson and beta-binomial).

Value

A numeric vector of the random samples.

See also

Examples

ran_student(10, theta = 1/2)
#>  [1]  0.7192964 -0.8942649  0.4530337 -0.4829191 -2.0921134 -0.3993548
#>  [7] 23.8070214  0.3146970 -0.1644658 -0.1594671